+++ src/main/java/kris/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package kris.command;
+
+import kris.TaskList;
+import kris.Ui;
+import kris.Storage;
+
 public class ByeCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
