+++ src/main/java/kris/command/DeadlineCommand.java
@@ -1,13 +1,22 @@
-public class EventCommand extends Command {
+package kris.command;
+
+import kris.TaskList;
+import kris.Ui;
+import kris.Storage;
+import kris.Parser;
+import kris.task.Deadline;
+import kris.exception.KrisException;
+
+public class DeadlineCommand extends Command {
     private String input;
     
-    public EventCommand(String input) {
+    public DeadlineCommand(String input) {
         this.input = input;
     }
     
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws KrisException {
-        Event newTask = Parser.parseEvent(input);
+        Deadline newTask = Parser.parseDeadline(input);
         tasks.add(newTask);
         ui.showTaskAdded(newTask, tasks.size());
         storage.save(tasks.getTasks());
