+++ src/main/java/kris/command/EventCommand.java
@@ -1,13 +1,22 @@
-public class DeadlineCommand extends Command {
+package kris.command;
+
+import kris.TaskList;
+import kris.Ui;
+import kris.Storage;
+import kris.Parser;
+import kris.task.Event;
+import kris.exception.KrisException;
+
+public class EventCommand extends Command {
     private String input;
     
-    public DeadlineCommand(String input) {
+    public EventCommand(String input) {
         this.input = input;
     }
     
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws KrisException {
-        Deadline newTask = Parser.parseDeadline(input);
+        Event newTask = Parser.parseEvent(input);
         tasks.add(newTask);
         ui.showTaskAdded(newTask, tasks.size());
         storage.save(tasks.getTasks());
