+++ src/main/java/kris/exception/EmptyDescriptionException.java
@@ -1,3 +1,5 @@
+package kris.exception;
+
 public class EmptyDescriptionException extends KrisException {
     public EmptyDescriptionException(String taskType) {
         super("Yo! The description of a " + taskType + " cannot be empty, my friend!");
