+++ src/main/java/kris/exception/InvalidCommandException.java
@@ -1,3 +1,5 @@
+package kris.exception;
+
 public class InvalidCommandException extends KrisException {
     public InvalidCommandException(String command) {
         super("Yo! I don't know what '" + command + "' means, my friend! Check out the help below.");
