+++ src/main/java/kris/exception/KrisException.java
@@ -1,3 +1,5 @@
+package kris.exception;
+
 public class KrisException extends Exception {
     public KrisException(String message) {
         super(message);
