+++ src/main/java/kris/task/Task.java
@@ -1,3 +1,5 @@
+package kris.task;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
