+++ src/main/java/kris/task/Todo.java
@@ -1,3 +1,5 @@
+package kris.task;
+
 public class Todo extends Task {
 
     public Todo(String description) {
